fix roads table distance data type bug

+++ database/migrations/2017_12_24_220542_create_roads_table.php
@@ -15,7 +15,7 @@ class CreateRoadsTable extends Migration
     {
         Schema::create('roads', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('distance');
+            $table->double('distance', 15,2);
             $table->timestamps();
         });
     }
